+++ modules/Shmuelie.Windows/Shmuelie.Windows.psm1
@@ -4,7 +4,8 @@ foreach ($script in Get-ChildItem $publicRoot -Filter '*.ps1' | Sort-Object Name
 }
 
 Export-ModuleMember -Function @(
-    'Get-InstalledApplications', 'Get-ServiceProcess',
+    'Get-AppInstallerApp', 'Get-InstalledApplications', 'Get-ServiceProcess',
     'Get-WindowsTerminalSettings', 'Get-WindowsTerminalProfile',
-    'Start-WindowsPerformanceRecorder', 'Stop-WindowsPerformanceRecorder'
+    'Start-WindowsPerformanceRecorder', 'Stop-WindowsPerformanceRecorder',
+    'Update-AppInstallerApp'
 )